Update: Multiple Textures

+++ Engine/Vertex.h
@@ -5,10 +5,10 @@ struct Vertex
 {
     int position[3];
     int normal[3];
-    int texCoord[2];
+    float texCoord[2];
 
     Vertex() = default; // Default constructor
-    Vertex(int pos[3], int norm[3], int tex[2])
+    Vertex(int pos[3], int norm[3], float tex[2])
     {
         memcpy(position, pos, sizeof(position));
         memcpy(normal, norm, sizeof(normal));
